# challenge-3/application/usecases/book/structures.go
package book

import (
	domainBook "secure/challenge-3/domain/book"
)

// NewBook is a struct that contains the data for a new book
type NewBook struct {
	Title       string `json:"title" example:"book title"`
	UserID      int    `json:"user_id" gorm:"index"`
	Description string `json:"description" example:"book description"`
}

// PaginationResultBook is a struct that contains the pagination result for book
type PaginationResultBook struct {
	Data       *[]domainBook.Book
	Total      int64
	Limit      int64
	Current    int64
	NextCursor uint
	PrevCursor uint
	NumPages   int64
}